[fact] Reorder and rename subpackages #29
