+++ src/main/java/se/kth/spork/spoon/RoledValue.java
@@ -1,4 +1,4 @@
-package se.kth.spork.merge.spoon;
+package se.kth.spork.spoon;
 
 import spoon.reflect.path.CtRole;
