+++ src/main/java/se/kth/spork/spoon/SpoonNode.java
@@ -1,6 +1,6 @@
-package se.kth.spork.merge.spoon;
+package se.kth.spork.spoon;
 
-import se.kth.spork.merge.ListNode;
+import se.kth.spork.base3dm.ListNode;
 import spoon.reflect.declaration.CtElement;
 
 public interface SpoonNode extends ListNode {
